refactor: member 테스트코드 일괄 수정
