+++ app/src/test/java/com/hotelJava/member/application/port/in/SignUpUseCaseTest.java
@@ -1,10 +1,11 @@
-package com.hotelJava.member.application.port.service;
+package com.hotelJava.member.application.port.in;
 
-import static org.mockito.Mockito.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.doReturn;
 
-import com.hotelJava.TestFixture;
+import com.hotelJava.CommandTestFixture;
 import com.hotelJava.common.error.exception.BadRequestException;
-import com.hotelJava.member.application.port.in.SignUpUseCase;
 import com.hotelJava.member.application.port.in.command.MemberSignUpCommand;
 import com.hotelJava.member.application.port.out.persistence.CheckDuplicatedMemberEmailPort;
 import com.hotelJava.member.application.port.out.persistence.FindMemberPort;
@@ -15,28 +16,31 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.SpyBean;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
-class MemberSignUpServiceTest {
-
+@Transactional
+class SignUpUseCaseTest {
   @Autowired private SignUpUseCase sut;
   @Autowired private FindMemberPort findMemberPort;
+  @SpyBean private PasswordEncoder passwordEncoder;
   @SpyBean private CheckDuplicatedMemberEmailPort memberEmailDuplicateCheckPort;
 
   @Test
-  @DisplayName("같은 이메일을 사용하는 회원이 없다면 회원가입에 성공한다.")
-  @Transactional
+  @DisplayName("같은 이메일을 사용하는 회원이 없다면 회원가입에 성공하며 비밀번호는 암호화된다.")
   void 회원가입성공() {
     // given
-    MemberSignUpCommand signUpCommand = TestFixture.getMemberSignUpCommand();
+    MemberSignUpCommand signUpCommand = CommandTestFixture.memberSignUpCommand();
 
     // when
+    doReturn("encryptedPassword").when(passwordEncoder).encode(anyString());
     Member registeredMember = sut.signUp(signUpCommand);
     Member findMember = findMemberPort.findByEmail(signUpCommand.getEmail());
 
     // then
-    Assertions.assertThat(registeredMember).isEqualTo(findMember);
+    assertThat(registeredMember).isEqualTo(findMember);
+    Assertions.assertThat(findMember.getPassword()).isEqualTo("encryptedPassword");
   }
 
   @Test
@@ -47,7 +51,7 @@ class MemberSignUpServiceTest {
     doReturn(true).when(memberEmailDuplicateCheckPort).isDuplicated(anyString());
 
     // when, then
-    Assertions.assertThatThrownBy(() -> sut.signUp(TestFixture.getMemberSignUpCommand()))
+    Assertions.assertThatThrownBy(() -> sut.signUp(CommandTestFixture.memberSignUpCommand()))
         .isInstanceOf(BadRequestException.class);
   }
 }
